tests/wpt/tests/css/css-inline/text-box-trim/text-box-on-non-root-inline-box.html

<!DOCTYPE html>
<title>text-box on non-root inline box does not affect layout bounds only the box's size.</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#text-box-trim">
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#text-box-edge">
<style>
@import "support/MetricsTestFont.css";

.test {
  font: 20px/1 MetricsTestFont;
  position: relative;
}
.aqua {
  background-color: aqua;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body>
<div class="test">
  <span class=aqua style="text-box: ex alphabetic;" data-offset-y=12 data-expected-height=4>ApÉx</span>
  <span class=aqua style="text-box: trim-start cap alphabetic;" data-offset-y=4 data-expected-height=16>ApÉx</span>
  <span class=aqua style="text-box: trim-end text alphabetic;" data-offset-y=0 data-expected-height=16>ApÉx</span>
</div>
<div class="test aqua" data-expected-height=30>
  <span style="font-size: 30px; text-box: ex alphabetic;">ApÉx</span>
</div>
<script>
  document.fonts.ready.then(() => checkLayout('.aqua'));
</script>
</body>
