third_party/blink/web_tests/platform/linux/external/wpt/css/css-inline/text-box-trim/text-box-on-non-root-inline-box-expected.txt

This is a testharness.js-based test.
[FAIL] .aqua 1
  assert_equals: \n<span class="aqua" style="text-box: ex alphabetic;" data-offset-y="16" data-expected-height="2">ApÉx</span>\nheight expected 2 but got 15
[FAIL] .aqua 3
  assert_equals: \n<span class="aqua" style="text-box: trim-end text alphabetic;" data-offset-y="8" data-expected-height="10">ApÉx</span>\nheight expected 10 but got 15
[FAIL] .aqua 4
  assert_equals: \n<div class="aqua" data-expected-height="18">\n  <span style="font-size: 18px; text-box: ex alphabetic;">ApÉx</span>\n</div>\nheight expected 18 but got 21
Harness: the test ran to completion.
